+++ CHANGES.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+## master
+- added pip
+- added jupyer and firefox
+
 ## v1.0.0
 - Creation of VM is automated by a vagrant script, and a `first_run.sh` script to execute after the VM is up for the first time
 - `update_VM.sh` will now by default update to the latest release (of the SuperBuild). Options have been added to modify this.